updated float First.cpp

+++ First.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main()
 {
   int a=5, b=2;
-  int c = a/b;
-  cout<<"The division is "<<c;
+  float c = a/b;
+  cout<<"The division is "<<(float)c;
   return 0;
 }
